+++ graphbrain/cli.py
@@ -29,6 +29,7 @@ def wrapper(fun, command=False, text=None):
     parser.add_argument('--lang', type=str, help='language', default='en')
     parser.add_argument('--pattern', type=str, help='edge pattern',
                         default='*')
+    parser.add_argument('--agent', type=str, help='agent name')
 
     args = parser.parse_args()